requirements.txt
streamlit
pandas
geopandas
folium
streamlit-folium
textblob
fbprophet